Add files via upload
This is my Portfolio Project that summarizes my Major projects done throughout my Master's in Science degree with a concentration in Information technology. It comprises of the details of the Project, Evidence of my work and the accomplishments of those projects. It is all majorly related to my Field of Interest in technical areas like SQL, NoSQL, Python for Exploratory Data Analysis and visualizations and a Tableau dashboard with meaningful insights about the IPL (Indian Premiere League) ODI Data Analyzed.