Add files via upload
The Connections File is my Python Project, a Short analysis using Python to check my LinkedIn connections over time, To view in detail about my connections so far, which industries I am connected to and with the analysis it showed me to which particular role industry professionals I should more network. This information is given with the help of charts and data as well in my python project. It is just a short analysis and further statistical analysis can be done!!